Fix: fix cirular import by reoganizing modules
