+++ src/careamics/config/data/ng_data_config.py
@@ -22,10 +22,8 @@ from pydantic import (
     model_validator,
 )
 
-from careamics.dataset_ng.image_stack.image_utils import are_czi_axes_valid
-
 from ..transformations import XYFlipConfig, XYRandomRotate90Config
-from ..validators import check_axes_validity
+from ..validators import check_axes_validity, check_czi_axes_validity
 from .patch_filter import (
     MaskFilterConfig,
     MaxFilterConfig,
@@ -124,11 +122,11 @@ class NGDataConfig(BaseModel):
     )
 
     # Dataset configuration
-    axes: str
-    """Axes of the data, as defined in SupportedAxes."""
-
     data_type: Literal["array", "tiff", "zarr", "czi", "custom"]
     """Type of input data."""
+
+    axes: str
+    """Axes of the data, as defined in SupportedAxes."""
 
     patching: PatchingStrategies = Field(..., discriminator="name")
     """Patching strategy to use. Note that `random` is the only supported strategy for
@@ -219,17 +217,16 @@ class NGDataConfig(BaseModel):
         ValueError
             If axes are not valid.
         """
-        # Validate axes
-        check_axes_validity(axes)
-
         # Additional validation for CZI files
         if info.data["data_type"] == "czi":
-            if not are_czi_axes_valid(axes):
+            if not check_czi_axes_validity(axes):
                 raise ValueError(
                     f"Provided axes '{axes}' are not valid. Axes must be in the "
                     f"`SC(Z/T)YX` format, where Z or T are optional, and S and C can be"
                     f" singleton dimensions, but must be provided."
                 )
+        else:
+            check_axes_validity(axes)
 
         return axes
