+++ src/careamics/config/validators/__init__.py
@@ -2,15 +2,17 @@
 
 __all__ = [
     "check_axes_validity",
+    "check_czi_axes_validity",
     "model_matching_in_out_channels",
     "model_without_final_activation",
     "model_without_n2v2",
     "patch_size_ge_than_8_power_of_2",
 ]
 
+from .axes_validators import check_axes_validity, check_czi_axes_validity
 from .model_validators import (
     model_matching_in_out_channels,
     model_without_final_activation,
     model_without_n2v2,
 )
-from .validator_utils import check_axes_validity, patch_size_ge_than_8_power_of_2
+from .patch_validators import patch_size_ge_than_8_power_of_2
