+++ src/careamics/dataset_ng/image_stack/image_utils/__init__.py
@@ -1,13 +1,11 @@
 """Image stack utilities."""
 
 __all__ = [
-    "are_czi_axes_valid",
     "pad_patch",
     "reshaped_array_shape",
 ]
 
 
-from .czi_image_stack_utils import are_czi_axes_valid
 from .image_stack_utils import (
     pad_patch,
     reshaped_array_shape,
