+++ tests/config/validators/test_validator_utils.py
@@ -1,51 +1,8 @@
 import pytest
 
 from careamics.config.validators import (
-    check_axes_validity,
     patch_size_ge_than_8_power_of_2,
 )
-
-
-@pytest.mark.parametrize(
-    "axes, valid",
-    [
-        # Passing
-        ("yx", True),
-        ("Yx", True),
-        ("Zyx", True),
-        ("STYX", True),
-        ("CYX", True),
-        ("YXC", True),
-        ("TzYX", True),
-        ("SZYX", True),
-        ("STZYX", True),
-        ("XY", True),
-        ("YXT", True),
-        ("ZTYX", True),
-        # non consecutive XY
-        ("YZX", False),
-        ("YZCXT", False),
-        # too few axes
-        ("", False),
-        ("X", False),
-        # no yx axes
-        ("ZT", False),
-        ("ZY", False),
-        # repeating characters
-        ("YYX", False),
-        ("YXY", False),
-        # invalid characters
-        ("YXm", False),
-        ("1YX", False),
-    ],
-)
-def test_are_axes_valid(axes, valid):
-    """Test if axes are valid"""
-    if valid:
-        check_axes_validity(axes)
-    else:
-        with pytest.raises((ValueError, NotImplementedError)):
-            check_axes_validity(axes)
 
 
 @pytest.mark.parametrize(
